all: normalize Go modules
Make all Go projects their own module, as they are completely unrelated
and might require different versions of the same dependency, assign the
ISC license to everything that should have one, and add package docs.

Finally, store in the attic replaybench and tlsbench. The idea was good,
but the execution wasn't.